move test in the right file after backport
